fix: recompute ulaPal on reset and flash register writes
reset() and flash register writes (regs 8/9) did not call
_recomputeUlaPal, leaving stale colours in the rendered palette.

Also consolidates test-nula.js into test-video.js per project
convention (one test file per component).
